STL-622: PR fix

+++ database/lib/src/util/database_migration_statement_factory.dart
@@ -59,7 +59,7 @@ abstract final class DatabaseMigrationStatementFactory {
   static String addRealColumnStatement({
     required String tableName,
     required String columnName,
-    required int defaultValue,
+    required double defaultValue,
   }) {
     return _addRealColumn(
       tableName: tableName,
@@ -100,7 +100,7 @@ abstract final class DatabaseMigrationStatementFactory {
     required String tableName,
     required String columnName,
     required bool isNullable,
-    int? defaultValue,
+    double? defaultValue,
   }) {
     return addColumnStatement(
       tableName: tableName,
